Refactroing

+++ vi_timing/test/source.c
@@ -6,6 +6,7 @@
 #include "../timing.h"
 
 #include <inttypes.h>
+#include <stdint.h>
 #include <threads.h>
 
 //VI_OPTIMIZE_OFF
@@ -31,7 +32,7 @@ void foo_c(void)
 {
 	vi_tmWarming(0, 100);
 	
-	printf("Version: \'%s\' [%"PRIdPTR"]\n", (const char*)vi_tmInfo(VI_TM_INFO_VERSION), vi_tmInfo(VI_TM_INFO_VER));
+	printf("Version: \'%s\' [%" PRIdPTR "]\n", (const char*)vi_tmInfo(VI_TM_INFO_VERSION), (intptr_t)vi_tmInfo(VI_TM_INFO_VER));
 
 	thrd_yield();
 	thrd_sleep(&(struct timespec) { .tv_nsec = 100 }, NULL);
